moved context and types to its own folders
