+++ client/src/components/Navbar.tsx
@@ -9,7 +9,7 @@ import Typography from '@material-ui/core/Typography';
 import { StylesProvider, useTheme } from '@material-ui/styles'; // make styled components styling have priority
 import styled from 'styled-components';
 
-import { AppContext } from '../AppContext';
+import { AppContext } from '../context/AppContext';
 import CSESocLogo from '../assets/notangles_one_n_with_grey.png';
 import CSESocLogoTwo from '../assets/notangles_two_n_with_grey.gif';
 import { ThemeType } from '../constants/theme';
