+++ client/src/index.tsx
@@ -2,12 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import ReactGA from 'react-ga';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+
 import App from './App';
-import AppContextProvider from './AppContext';
 import Privacy from './components/Privacy';
-import './index.css';
-import CourseContextProvider from './CourseContext';
+import AppContextProvider from './context/AppContext';
+import CourseContextProvider from './context/CourseContext';
 import * as serviceWorker from './serviceWorker';
+
+import './index.css';
 
 const GOOGLE_ANALYTICS_ID = process.env.REACT_APP_GOOGLE_ANALYTICS_ID;
